Revert "Merge branch 'master' into cyberiad-med-rnd-remap"
This reverts commit 93cb46dfb0906d708fbfeaeb3045d719a8ebfd60, reversing
changes made to 1c0a4087ee6a339dba717b92a40653e81da7a0c1.

+++ code/modules/mob/living/inhand_holder.dm
@@ -6,12 +6,10 @@
 	icon = null
 	icon_state = ""
 	slot_flags = NONE
-	/// Mob inside of us
 	var/mob/living/held_mob
-	/// True if we've started being destroyed
 	var/destroying = FALSE
 
-/obj/item/clothing/head/mob_holder/Initialize(mapload, mob/living/held_mob, worn_state, head_icon, lh_icon, rh_icon, worn_slot_flags = NONE)
+/obj/item/clothing/head/mob_holder/Initialize(mapload, mob/living/M, worn_state, head_icon, lh_icon, rh_icon, worn_slot_flags = NONE)
 	if(head_icon)
 		worn_icon = head_icon
 	if(worn_state)
@@ -22,38 +20,29 @@
 		righthand_file = rh_icon
 	if(worn_slot_flags)
 		slot_flags = worn_slot_flags
-	update_weight_class(held_mob.held_w_class)
-	insert_mob(held_mob)
-	return ..()
+	update_weight_class(M.held_w_class)
+	deposit(M)
+	. = ..()
 
 /obj/item/clothing/head/mob_holder/Destroy()
 	destroying = TRUE
 	if(held_mob)
-		release()
+		release(FALSE)
 	return ..()
 
-/obj/item/clothing/head/mob_holder/proc/insert_mob(mob/living/new_prisoner)
-	if(!istype(new_prisoner))
+/obj/item/clothing/head/mob_holder/proc/deposit(mob/living/L)
+	if(!istype(L))
 		return FALSE
-	new_prisoner.setDir(SOUTH)
-	update_visuals(new_prisoner)
-	held_mob = new_prisoner
-	RegisterSignal(held_mob, COMSIG_QDELETING, PROC_REF(on_mob_deleted))
-	new_prisoner.forceMove(src)
-	name = new_prisoner.name
-	desc = new_prisoner.desc
+	L.setDir(SOUTH)
+	update_visuals(L)
+	held_mob = L
+	L.forceMove(src)
+	name = L.name
+	desc = L.desc
 	return TRUE
 
-/obj/item/clothing/head/mob_holder/proc/on_mob_deleted()
-	SIGNAL_HANDLER
-	held_mob = null
-	if (isliving(loc))
-		var/mob/living/holder = loc
-		holder.temporarilyRemoveItemFromInventory(src, force = TRUE)
-	qdel(src)
-
-/obj/item/clothing/head/mob_holder/proc/update_visuals(mob/living/held_guy)
-	appearance = held_guy.appearance
+/obj/item/clothing/head/mob_holder/proc/update_visuals(mob/living/L)
+	appearance = L.appearance
 
 /obj/item/clothing/head/mob_holder/on_thrown(mob/living/carbon/user, atom/target)
 	if((item_flags & ABSTRACT) || HAS_TRAIT(src, TRAIT_NODROP))
@@ -67,29 +56,29 @@
 	release()
 	return throw_mob
 
-/obj/item/clothing/head/mob_holder/Moved(atom/old_loc, movement_dir, forced, list/old_locs, momentum_change)
-	. = ..()
+/obj/item/clothing/head/mob_holder/dropped()
+	..()
 	if(held_mob && isturf(loc))
 		release()
 
-/obj/item/clothing/head/mob_holder/proc/release(display_messages = TRUE)
+/obj/item/clothing/head/mob_holder/proc/release(del_on_release = TRUE, display_messages = TRUE)
 	if(!held_mob)
-		if(!destroying)
+		if(del_on_release && !destroying)
 			qdel(src)
 		return FALSE
 	var/mob/living/released_mob = held_mob
 	held_mob = null // stops the held mob from being release()'d twice.
 	if(isliving(loc))
-		var/mob/living/captor = loc
+		var/mob/living/L = loc
 		if(display_messages)
-			to_chat(captor, span_warning("[released_mob.declent_ru(NOMINATIVE)] вырывается из рук!"))
-		captor.dropItemToGround(src)
+			to_chat(L, span_warning("[capitalize(released_mob.declent_ru(NOMINATIVE))] вырывается из рук!"))
+		L.dropItemToGround(src)
 	released_mob.forceMove(drop_location())
 	released_mob.reset_perspective()
 	released_mob.setDir(SOUTH)
 	if(display_messages)
-		released_mob.visible_message(span_warning("[released_mob.declent_ru(NOMINATIVE)] расправляется!"))
-	if(!destroying)
+		released_mob.visible_message(span_warning("[capitalize(released_mob.declent_ru(NOMINATIVE))] расправляется!"))
+	if(del_on_release && !destroying)
 		qdel(src)
 	return TRUE
 
@@ -101,14 +90,14 @@
 
 /obj/item/clothing/head/mob_holder/Exited(atom/movable/gone, direction)
 	. = ..()
-	if(held_mob == gone)
+	if(held_mob && held_mob == gone)
 		release()
 
 /obj/item/clothing/head/mob_holder/on_found(mob/finder)
 	if(held_mob?.will_escape_storage())
-		to_chat(finder, span_warning("[held_mob.declent_ru(NOMINATIVE)] вылезает! "))
-		finder.visible_message(span_warning("[held_mob.declent_ru(NOMINATIVE)] вылезает из контейнера, который открывает [finder.declent_ru(NOMINATIVE)]!"), ignored_mobs = finder)
-		release(display_messages = FALSE)
+		to_chat(finder, span_warning("[capitalize(held_mob.declent_ru(NOMINATIVE))] вылезает! "))
+		finder.visible_message(span_warning("[capitalize(held_mob.declent_ru(NOMINATIVE))] вылезает из контейнера, который открывает [finder.declent_ru(NOMINATIVE)]!"), ignored_mobs = finder)
+		release(TRUE, FALSE)
 		return
 
 /obj/item/clothing/head/mob_holder/drone/Initialize(mapload, mob/living/M, worn_state, head_icon, lh_icon, rh_icon, worn_slot_flags = NONE)
@@ -117,16 +106,15 @@
 		return INITIALIZE_HINT_QDEL
 	return ..()
 
-/obj/item/clothing/head/mob_holder/drone/insert_mob(mob/living/new_prisoner)
+/obj/item/clothing/head/mob_holder/drone/deposit(mob/living/L)
 	. = ..()
-	if(!isdrone(new_prisoner))
+	if(!isdrone(L))
 		qdel(src)
-		return
 	name = "drone (hiding)"
 	desc = "This drone is scared and has curled up into a ball!"
 
-/obj/item/clothing/head/mob_holder/drone/update_visuals(mob/living/contained)
-	var/mob/living/basic/drone/drone = contained
+/obj/item/clothing/head/mob_holder/drone/update_visuals(mob/living/L)
+	var/mob/living/basic/drone/drone = L
 	if(!drone)
 		return ..()
 	icon = 'icons/mob/silicon/drone.dmi'
@@ -136,10 +124,10 @@
 
 /obj/item/clothing/head/mob_holder/destructible/Destroy()
 	if(held_mob)
-		release(display_messages = TRUE, delete_mob = TRUE)
+		release(FALSE, TRUE, TRUE)
 	return ..()
 
-/obj/item/clothing/head/mob_holder/destructible/release(display_messages = TRUE, delete_mob = FALSE)
+/obj/item/clothing/head/mob_holder/destructible/release(del_on_release = TRUE, display_messages = TRUE, delete_mob = FALSE)
 	if(delete_mob && held_mob)
 		QDEL_NULL(held_mob)
 	return ..()
